+++ CHANGELOG.md
@@ -22,6 +22,24 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 - Mix formatter configuration for using `defvocab` without parens 
 
 
+### Changed
+
+- `RDF.Serialization.Writer.write_file/4` which is the basis used by all the
+  `write_file/3` and `write_file!/3` functions of all serialization format modules
+  like `RDF.NTriples`, `RDF.Turtle`, `JSON.LD` etc. now opens file in a different 
+  mode: it no longer opens them with the [`:utf8` option](https://hexdocs.pm/elixir/File.html#open/2).
+  First, this by default slowed down the writing, but more importantly could lead  
+  to unexpected encoding issues. 
+  This is a **breaking change**: If your code relied on this file mode, you can  
+  get the old behaviour, by specifying the `file_mode` on these functions 
+  accordingly as `[:utf8, :write, :exclusive]`. For example, to write a Turtle 
+  file with the old behaviour, you can do it like this:
+  
+```elixir
+RDF.Turtle.write_file!(some_data, some_path, file_mode: ~w[utf8 write exclusive]a)
+``` 
+
+
 [Compare v0.6.2...HEAD](https://github.com/marcelotto/rdf-ex/compare/v0.6.2...HEAD)